first version of a script to regress the sensitivities of N-1 participants onto the eventfiles of the localization data.
This is attempt one and needs a lot of changes. For now, it will compute one GLM per pair of ROIs and subject, a revised version will do only one glm for all subjects. A revised version will also have a more modular structure, this one is just a first brain barf.