Add one script to rule them all.
This script does the classification, glm, and plotting on the localizer and avmovie task, with either linear gnb, linear sgd, or 1-vs-all sgd, for stripped or full datasets, with, without and with only coordinates, for ROIs for each hemisphere separately or combined. Specifying the command line arguments will come close to writing a novel. Nothing is tested, I will find and kill bugs tomorrow ;-)