Add main analysis script as file (again).
Something I did in the past turned the original cv_clf_glm.py script into a ... subdataset? Idk for sure, but tracking of changes was not possible, it always committed the full script again. For the sake of time, I'm not fixing the old script and merely copy its content into this new file, and continue developing this file instead of the other one.